refactor: manager class in CMMS
